requirements.txt
zlibrary==1.0.2
flask==2.3.3
pandas==2.0.3
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.8.5
asyncio==3.4.3
jinja2==3.1.2
werkzeug==2.3.7